blog/requirements.txt
Django>=4.2,<5.0
mysqlclient>=2.2.0
Pillow>=10.0.0
python-slugify>=8.0.0
